Additional: structured header templates and render | FT-28: refactored relinkingApi
