+++ src/js/templates/header/createHeaderHomeMarkup.js
@@ -1,5 +1,5 @@
-<!-- <header class="header container">
-  <div class="nav-wrap">
+export default function createHeaderHomeMarkup() {
+  return `<div class="nav-wrap">
     <a href="#" class="header__logo" id="logoLink">
       <svg
         class="header__icon"
@@ -62,4 +62,5 @@
   <p class="notification">
     Search result is not successful. Enter the correct movie name and try again.
   </p>
-</header> -->
+`;
+}
